tambahan ica cantik yang cute semangat skripsinya

+++ database/migrations/2021_06_24_145301_create_tbl_kriteria.php
@@ -17,7 +17,7 @@ class CreateTblKriteria extends Migration
             $table->bigIncrements('id');
             $table->string('id_aspek');
             $table->string('nama_kriteria');
-            $table->float('nilai_ideal');
+            $table->integer('nilai_ideal');
             $table->timestamps();
         });
     }
